modules/predictor/requirements.txt
edge_impulse_linux
azure-iot-device
numpy
pyaudio
six
opencv-python-headless
psutil
